requirements.txt
fastapi
uvicorn
discord.py
pyjwt
httpx
beautifulsoup4
